+++ client/src/Containers/Access.js
@@ -1,0 +1,11 @@
+import { connect } from 'react-redux';
+import { requestAccess, grantAccess, clearError, joinWithCode } from '../store/actions';
+import PrivateAccessModal from '../Components/UI/Modal/PrivateAccess'
+
+const mapStateToProps = (store, ownProps) => ({
+  error: store.loading.errorMessage,
+  resource: store[ownProps.resource].byId[ownProps.id],
+  belongsToCourse: ownProps.course,
+})
+
+export default connect(mapStateToProps, { requestAccess, grantAccess, joinWithCode, clearError})(PrivateAccessModal)
