+++ client/src/Containers/index.js
@@ -8,3 +8,4 @@ export { default as MyVMT } from './MyVMT';
 export { default as Course } from './Course';
 export { default as Activity } from './Activity';
 export { default as Room } from './Room';
+export { default as Access } from './Access';
